Add Coursera Provided files
Coursera course provides support files for this project.

Note from Coursera:
The Persistence file contains a simple core data setup boilerplate code that sets up a simple database file in memory.

FetchedObjets file contains code that makes it easier to fetch the models from the database. As it was explained in the last course and is not the focus of this exercise, you don't need to worry about it. Just consider it as it is, or in other words, as a struct that retrieves objects from Core Data.