Add Search Functionality
Search Functionality added to the project, allowing user to search for a menu item.

A TextField with a Search State parameter is added.
A function is then created that returns an NSDescriptor containing the input into the TextField from earlier. This is added as an input parameter to the constructor for FetchedObjects.